SGA-12135 - eliminate the harder requirement of using lifecycle.ignore_changes
Eliminates the harder requirement of using lifecycle.ignore_changes for satori_auth_settings.0.credentials.0.password only.

Since the password is not returned from the backend, the Terraform state has no idea if it was changed or not.

Therefore, the lifecycle.ignore_changes was introduced, which ignores the change of this property.

However, it can be challenging for automation, as any password update requires 2 changes in tf file, first, without lifecycle.ignore_changes and the second with.

The fix introduces the manipulation of Terraform state, which eliminates the harder requirement of using lifecycle.ignore_changes